chore(gatsby-telemetry): Migrate flush to TypeScript
