--- packages/gatsby-telemetry/src/flush.js
@@ -1,17 +1,0 @@
-const { join } = require(`path`)
-const { fork } = require(`child_process`)
-
-module.exports = isTrackingEnabled => async () => {
-  if (!isTrackingEnabled) {
-    return
-  }
-
-  // Submit events on background w/o blocking the main process
-  // nor relying on it's lifecycle
-  const forked = fork(join(__dirname, `send.js`), {
-    detached: true,
-    stdio: `ignore`,
-    execArgv: [],
-  })
-  forked.unref()
-}
